Added glossary button
